+++ middleware.js
@@ -1,3 +1,7 @@
+const {campgroundSchema,reviewSchema} = require('./schemas.js');
+const ExpressError = require('./utils/ExpressError');
+const Campground = require("./models/campground");
+
 module.exports.isLoggedIn= (req, res, next)=>{
     if(!req.isAuthenticated()){
         //store the url the user was trying to hit
@@ -6,4 +10,36 @@ module.exports.isLoggedIn= (req, res, next)=>{
         return res.redirect('/login')
     }
     next();
+}
+
+//validateCampground middleware
+module.exports.validateCampground = (req, res, next) =>{
+    const {error} = campgroundSchema.validate(req.body);
+    if(error){
+        const msg = error.details.map(el => el.message).join(',');
+        throw new ExpressError(msg, 400);
+    }else{
+        next();
+    }
+}
+
+//middleware for author authorization
+module.exports.isAuthor = async(req, res, next)=>{
+    const {id}= req.params;
+    const campground = await Campground.findById(id);
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error','You do not have permission to do that');
+        return res.redirect(`/campgrounds/${campground._id}`);
+    }
+    next();
+}
+
+module.exports.validateReview = (req, res, next)=> {
+    const {error} = reviewSchema.validate(req.body);
+    if(error){
+        const msg = error.details.map(el => el.message).join(',')
+        throw new ExpressError(msg, 400);
+    }else{
+        next();
+    }
 }
