+++ routes/campground.js
@@ -1,20 +1,13 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require("../utils/catchAsync");
+
 const ExpressError = require('../utils/ExpressError');
 const Campground = require("../models/campground");
-const {campgroundSchema} = require('../schemas.js');
-const {isLoggedIn} = require('../middleware')
 
-const validateCampground = (req, res, next) =>{
-    const {error} = campgroundSchema.validate(req.body);
-    if(error){
-        const msg = error.details.map(el => el.message).join(',');
-        throw new ExpressError(msg, 400);
-    }else{
-        next();
-    }
-}
+const {isLoggedIn, isAuthor, validateCampground} = require('../middleware');
+const campground = require('../models/campground');
+
 
 router.get("/", catchAsync(async(req, res)=>{
     const campgrounds = await Campground.find({});
@@ -31,14 +24,15 @@ router.post('/',isLoggedIn, validateCampground, catchAsync(async(req, res, next)
     if(!req.body.campground) throw new ExpressError('Invalid Campground Data', 400)
     
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash('success','Successfully made a new campground!');
     res.redirect(`/campgrounds/${campground._id}`);
 }))
 
 router.get("/:id",isLoggedIn, catchAsync(async(req, res)=>{
-    const campground = await Campground.findById(req.params.id).populate('reviews');
-    //console.log(campground)
+    const campground = await Campground.findById(req.params.id).populate('reviews').populate('author');
+    // console.log(campground)
     if(!campground){
         req.flash('error', 'cannot find that campground');
         return res.redirect('/campgrounds');
@@ -46,23 +40,24 @@ router.get("/:id",isLoggedIn, catchAsync(async(req, res)=>{
     res.render('campgrounds/show',{campground});
 }))
 
-router.get('/:id/edit',isLoggedIn, catchAsync(async(req, res)=>{
+router.get('/:id/edit',isLoggedIn, isAuthor, catchAsync(async(req, res)=>{
     const campground = await Campground.findById(req.params.id);
-    if(!campground){
+    if(!campground){ 
         req.flash('error', 'sorry cannot find that campground!');
         return res.redirect('/campgrounds');
     }
     res.render('campgrounds/edit',{campground});
 }))
 
-router.put('/:id', isLoggedIn, validateCampground, catchAsync(async(req, res)=>{
-    const {id}= req.params;
+router.put('/:id', isLoggedIn, isAuthor, validateCampground, catchAsync(async(req, res)=>{
+    const{id}= req.params;
     const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground})
     req.flash('success', 'Successfully updated campground');
     res.redirect(`/campgrounds/${campground._id}`);
 }))
 
-router.delete('/:id',isLoggedIn, catchAsync(async(req, res)=>{
+router.delete('/:id',isLoggedIn,isAuthor, catchAsync(async(req, res)=>{
+    
     const {id} = req.params;
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted campground');
