+++ routes/review.js
@@ -3,20 +3,10 @@ const router = express.Router({mergeParams: true});
 
 const Campground = require('../models/campground')
 const Review = require("../models/review");
-const {reviewSchema} = require('../schemas.js');
+const{validateReview}= require('../middleware')
 const catchAsync = require("../utils/catchAsync");
 const ExpressError = require('../utils/ExpressError');
 const {isLoggedIn} = require('../middleware');
-
-const validateReview = (req, res, next)=> {
-    const {error} = reviewSchema.validate(req.body);
-    if(error){
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400);
-    }else{
-        next();
-    }
-}
 
 router.post('/', validateReview,isLoggedIn, catchAsync(async(req, res)=>{
     const campground = await Campground.findById(req.params.id);
